requirements.txt
requests
python-dotenv
pydantic
rich


pytest>=8.0.0
